Consolidate webpack configurations (#2716) bruception
* Consolidate webpack configurations

* Add lint step back into dev compilation
