+++ frontend/package.json
@@ -31,7 +31,7 @@
     "circular-dependency-plugin": "5.2.2",
     "dart-sass": "^1.25.0",
     "del": "^6.0.0",
-    "eslint-webpack-plugin": "^3.1.1",
+    "eslint-webpack-plugin": "3.1.1",
     "grecaptcha": "^1.0.3",
     "gulp": "^4.0.2",
     "gulp-concat": "^2.6.1",
@@ -47,7 +47,8 @@
     "vinyl-buffer": "^1.0.1",
     "vinyl-paths": "^3.0.1",
     "webpack": "^5.68.0",
-    "webpack-cli": "4.9.2"
+    "webpack-cli": "4.9.2",
+    "webpack-merge": "5.8.0"
   },
   "dependencies": {
     "axios": "^0.21.2",
